Ensure payloads are not removed before parsing in `put_message`
Currently, the put_message handler pops the payload off a message
prior to parsing it. However, it also does this before it tries
to parse the message, or store it. As such, if a message exceeds
500bytes, it will cause an error on message parse if the payload_digest
was not sent by the client. This commit moves the mutation until
after parsing, and after storing.